requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-MySQLdb==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.1
patsy==0.5.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
